+++ src/recoil/userInitInfoRecoil.tsx
@@ -4,3 +4,18 @@ export const userFeelingTimeZoneState = atom({
   key: 'userFeelingTimeZoneState',
   default: 0,
 });
+
+export const userNotificationTimeState = atom({
+  key: 'userNotificationTimeState',
+  default: 0,
+});
+
+export const userFeelingWeatherState = atom({
+  key: 'userFeelingWeatherState',
+  default: 'option1',
+});
+
+export const userNameState = atom({
+  key: 'userNameState',
+  default: '',
+});
