--- src/recoil/userNameRecoil.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const userNameState = atom({
-  key: 'userNameState',
-  default: '',
-});
