--- src/recoil/userWeather.tsx
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const userWeatherState = atom({
-  key: 'userWeatherState',
-  default: 'option1',
-});
